requirements.txt
fpdf2
opencv-python
tk
pillow
djitellopy
ultralytics
torch
pynput